changed rescale intensity to work on doubles as well as ints
git-svn-id: https://svn.physiomeproject.org/svn/cmiss/cmgui/trunk@5300 3c1deb5b-d424-0410-962d-aba41a686d42

+++ source/image_processing/computed_field_rescaleIntensityImageFilter.h
@@ -61,7 +61,7 @@ DESCRIPTION :
 ==============================================================================*/
 
 int Computed_field_set_type_rescale_intensity_image_filter(struct Computed_field *field,
-	struct Computed_field *source_field, int outputMin, int outputMax);
+	struct Computed_field *source_field, double outputMin, double outputMax);
 /*******************************************************************************
 LAST MODIFIED : 18 Nov 2006
 
@@ -72,7 +72,7 @@ Sets number of components to same number as <source_field>.
 ==============================================================================*/
 
 int Computed_field_get_type_rescale_intensity_image_filter(struct Computed_field *field,
-	struct Computed_field **source_field, int *outputMin, int *outputMax);
+	struct Computed_field **source_field, double *outputMin, double *outputMax);
 /*******************************************************************************
 LAST MODIFIED : 18 Nov 2006
